move and deduplicate civisibility tests
